✅ detect files
